refactor : 테스트 코드의 파일 위치 조정 및 Utils의 패키지명 변경
- InputUtils.java의 파일 위치를 main과 같이 조정
- Main의 Utils 패키지 명을 옳바르게 소문자로 변경
