+++ src/main/java/utils/PrintUtils.java
@@ -1,4 +1,4 @@
-package Utils;
+package utils;
 
 import racingcar.Car;
